feat(rank): a League-style rank with LP, and an account level from XP
Splits the one XP number in two (ADR 0007):

- Rank: rank points in divisions of 100 LP, Mingo I to the top. Each answer
  moves it Elo-style against the question's rating, so blind guessing never
  pays; daily and mock count most, practice little, live quizzes not at all.
  Stakes grow by division; after three wrong in a row losses halve and the
  next right answer pays 1.5x. "I'm not sure" costs half a wrong answer.
  Positions place you (no floor); every 1 September a soft reset. Training
  wheels follow the division and come back if you drop.
- Account level: XP only goes up. Every answer earns some; right ones get
  first-win (+50 %, first three a day), combo, streak and rare critical
  bonuses, added. Fast early levels, emblem frames at 10/25/50/100.
- Leaderboard: ranked by rank this season; the week and areas by LP won;
  verticals by average rank. Captains by rank.
- Seeded pacing simulation as a unit test; migration 0014 backfills.

+++ tests/unit/test_leaderboard_rules.py
@@ -76,7 +76,7 @@ def test_periods_start_at_midnight_in_madrid(now: datetime, period: str, since: 
     assert madrid_midnight(first_day(period, madrid_day(now))) == since
 
 
-def m(vertical: str | None, points: int, played: bool = False) -> Member:
+def m(vertical: str | None, points: float, played: bool = False) -> Member:
     return Member(vertical, points, played)
